refactor(engine-core): Centralizing diffing algo logic (#2641)
