+++ packages/@lwc/engine-core/src/framework/hooks.ts
@@ -11,20 +11,50 @@ import {
     assert,
     create,
     isArray,
+    isTrue,
     isFalse,
     isNull,
     isUndefined,
     keys,
+    SVG_NAMESPACE,
+    KEY__SHADOW_RESOLVER,
 } from '@lwc/shared';
-import { getClassList, setText, getAttribute, remove, insert } from '../renderer';
-import { EmptyArray, parseStyleText } from './utils';
-import { createVM, getAssociatedVMIfPresent, VM, ShadowMode, RenderMode } from './vm';
-import { VNode, VCustomElement, VElement, VNodes } from '../3rdparty/snabbdom/types';
-import { updateDynamicChildren, updateStaticChildren } from '../3rdparty/snabbdom/snabbdom';
+
+import {
+    remove,
+    insert,
+    nextSibling,
+    createElement,
+    createText,
+    setText,
+    createComment,
+    getAttribute,
+    getClassList,
+    isSyntheticShadowDefined,
+} from '../renderer';
+import { logError, logWarn } from '../shared/logger';
+
+import {
+    createVM,
+    getRenderRoot,
+    appendVM,
+    removeVM,
+    hydrateVM,
+    rerenderVM,
+    getAssociatedVM,
+    getAssociatedVMIfPresent,
+    runConnectedCallback,
+    VM,
+    VMState,
+    ShadowMode,
+    RenderMode,
+} from './vm';
 import { patchElementWithRestrictions, unlockDomMutation, lockDomMutation } from './restrictions';
 import { getComponentInternalDef } from './def';
 import { markComponentAsDirty } from './component';
-import { logError } from '../shared/logger';
+import { getUpgradableConstructor } from './upgradable-element';
+import { EmptyArray, parseStyleText } from './utils';
+import { VNode, VNodes, VCustomElement, VElement, VText, VComment, Hooks, Key } from './vnodes';
 
 import { patchAttributes } from './modules/attrs';
 import { patchProps } from './modules/props';
@@ -33,6 +63,263 @@ import { patchStyleAttribute } from './modules/computed-style-attr';
 import { applyEventListeners } from './modules/events';
 import { applyStaticClassAttribute } from './modules/static-class-attr';
 import { applyStaticStyleAttribute } from './modules/static-style-attr';
+
+const enum LwcDomMode {
+    manual = 'manual',
+}
+
+export const TextHook: Hooks<VText> = {
+    create: (vnode) => {
+        const { owner } = vnode;
+
+        const elm = createText(vnode.text!);
+        linkNodeToShadow(elm, owner);
+        vnode.elm = elm;
+    },
+    update: updateNodeHook,
+    insert: insertNode,
+    move: insertNode, // same as insert for text nodes
+    remove: removeNode,
+    hydrate: (vNode: VNode, node: Node) => {
+        if (process.env.NODE_ENV !== 'production') {
+            // eslint-disable-next-line lwc-internal/no-global-node
+            if (node.nodeType !== Node.TEXT_NODE) {
+                logError('Hydration mismatch: incorrect node type received', vNode.owner);
+                assert.fail('Hydration mismatch: incorrect node type received.');
+            }
+
+            if (node.nodeValue !== vNode.text) {
+                logWarn(
+                    'Hydration mismatch: text values do not match, will recover from the difference',
+                    vNode.owner
+                );
+            }
+        }
+
+        // always set the text value to the one from the vnode.
+        node.nodeValue = vNode.text ?? null;
+        vNode.elm = node;
+    },
+};
+
+export const CommentHook: Hooks<VComment> = {
+    create: (vnode) => {
+        const { owner, text } = vnode;
+
+        const elm = createComment(text);
+        linkNodeToShadow(elm, owner);
+        vnode.elm = elm;
+    },
+    update: updateNodeHook,
+    insert: insertNode,
+    move: insertNode,
+    remove: removeNode,
+    hydrate: (vNode: VNode, node: Node) => {
+        if (process.env.NODE_ENV !== 'production') {
+            // eslint-disable-next-line lwc-internal/no-global-node
+            if (node.nodeType !== Node.COMMENT_NODE) {
+                logError('Hydration mismatch: incorrect node type received', vNode.owner);
+                assert.fail('Hydration mismatch: incorrect node type received.');
+            }
+
+            if (node.nodeValue !== vNode.text) {
+                logWarn(
+                    'Hydration mismatch: comment values do not match, will recover from the difference',
+                    vNode.owner
+                );
+            }
+        }
+
+        // always set the text value to the one from the vnode.
+        node.nodeValue = vNode.text ?? null;
+        vNode.elm = node;
+    },
+};
+
+// insert is called after update, which is used somewhere else (via a module)
+// to mark the vm as inserted, that means we cannot use update as the main channel
+// to rehydrate when dirty, because sometimes the element is not inserted just yet,
+// which breaks some invariants. For that reason, we have the following for any
+// Custom Element that is inserted via a template.
+export const ElementHook: Hooks<VElement> = {
+    create: (vnode) => {
+        const {
+            sel,
+            owner,
+            data: { svg },
+        } = vnode;
+
+        const namespace = isTrue(svg) ? SVG_NAMESPACE : undefined;
+        const elm = createElement(sel, namespace);
+
+        linkNodeToShadow(elm, owner);
+        fallbackElmHook(elm, vnode);
+        vnode.elm = elm;
+
+        patchElementPropsAndAttrs(null, vnode);
+    },
+    update: (oldVnode, vnode) => {
+        patchElementPropsAndAttrs(oldVnode, vnode);
+        patchChildren(vnode.elm!, oldVnode.children, vnode.children);
+    },
+    insert: (vnode, parentNode, referenceNode) => {
+        insertNode(vnode, parentNode, referenceNode);
+        createChildrenHook(vnode);
+    },
+    move: insertNode,
+    remove: (vnode, parentNode) => {
+        removeNode(vnode, parentNode);
+        removeChildren(vnode);
+    },
+    hydrate: (vnode, node) => {
+        const elm = node as Element;
+        vnode.elm = elm;
+
+        const { context } = vnode.data;
+        const isDomManual = Boolean(
+            !isUndefined(context) &&
+                !isUndefined(context.lwc) &&
+                context.lwc.dom === LwcDomMode.manual
+        );
+
+        if (isDomManual) {
+            // it may be that this element has lwc:inner-html, we need to diff and in case are the same,
+            // remove the innerHTML from props so it reuses the existing dom elements.
+            const { props } = vnode.data;
+            if (!isUndefined(props) && !isUndefined(props.innerHTML)) {
+                if (elm.innerHTML === props.innerHTML) {
+                    delete props.innerHTML;
+                } else {
+                    logWarn(
+                        `Mismatch hydrating element <${elm.tagName.toLowerCase()}>: innerHTML values do not match for element, will recover from the difference`,
+                        vnode.owner
+                    );
+                }
+            }
+        }
+
+        hydrateElmHook(vnode);
+
+        if (!isDomManual) {
+            hydrateChildrenHook(vnode.elm.childNodes, vnode.children, vnode.owner);
+        }
+    },
+};
+
+export const CustomElementHook: Hooks<VCustomElement> = {
+    create: (vnode) => {
+        const { sel, owner } = vnode;
+        const UpgradableConstructor = getUpgradableConstructor(sel);
+        /**
+         * Note: if the upgradable constructor does not expect, or throw when we new it
+         * with a callback as the first argument, we could implement a more advanced
+         * mechanism that only passes that argument if the constructor is known to be
+         * an upgradable custom element.
+         */
+        const elm = new UpgradableConstructor((elm: HTMLElement) => {
+            // the custom element from the registry is expecting an upgrade callback
+            createViewModelHook(elm, vnode);
+        });
+
+        linkNodeToShadow(elm, owner);
+        vnode.elm = elm;
+
+        const vm = getAssociatedVMIfPresent(elm);
+        if (vm) {
+            allocateChildrenHook(vnode, vm);
+        } else if (vnode.ctor !== UpgradableConstructor) {
+            throw new TypeError(`Incorrect Component Constructor`);
+        }
+        patchElementPropsAndAttrs(null, vnode);
+    },
+    update: (oldVnode, vnode) => {
+        patchElementPropsAndAttrs(oldVnode, vnode);
+        const vm = getAssociatedVMIfPresent(vnode.elm);
+        if (vm) {
+            // in fallback mode, the allocation will always set children to
+            // empty and delegate the real allocation to the slot elements
+            allocateChildrenHook(vnode, vm);
+        }
+        // in fallback mode, the children will be always empty, so, nothing
+        // will happen, but in native, it does allocate the light dom
+        patchChildren(vnode.elm!, oldVnode.children, vnode.children);
+        if (vm) {
+            if (process.env.NODE_ENV !== 'production') {
+                assert.isTrue(
+                    isArray(vnode.children),
+                    `Invalid vnode for a custom element, it must have children defined.`
+                );
+            }
+            // this will probably update the shadowRoot, but only if the vm is in a dirty state
+            // this is important to preserve the top to bottom synchronous rendering phase.
+            rerenderVM(vm);
+        }
+    },
+    insert: (vnode, parentNode, referenceNode) => {
+        insertNode(vnode, parentNode, referenceNode);
+        const vm = getAssociatedVMIfPresent(vnode.elm);
+        if (vm) {
+            if (process.env.NODE_ENV !== 'production') {
+                assert.isTrue(vm.state === VMState.created, `${vm} cannot be recycled.`);
+            }
+            runConnectedCallback(vm);
+        }
+        createChildrenHook(vnode);
+        if (vm) {
+            appendVM(vm);
+        }
+    },
+    move: insertNode,
+    remove: (vnode, parentNode) => {
+        removeNode(vnode, parentNode);
+        const vm = getAssociatedVMIfPresent(vnode.elm);
+        if (vm) {
+            // for custom elements we don't have to go recursively because the removeVM routine
+            // will take care of disconnecting any child VM attached to its shadow as well.
+            removeVM(vm);
+        }
+    },
+    hydrate: (vnode, elm) => {
+        // the element is created, but the vm is not
+        const { sel, mode, ctor, owner } = vnode;
+
+        const def = getComponentInternalDef(ctor);
+        createVM(elm, def, {
+            mode,
+            owner,
+            tagName: sel,
+        });
+
+        vnode.elm = elm as Element;
+
+        const vm = getAssociatedVM(elm);
+        allocateChildrenHook(vnode, vm);
+
+        hydrateElmHook(vnode);
+
+        // Insert hook section:
+        if (process.env.NODE_ENV !== 'production') {
+            assert.isTrue(vm.state === VMState.created, `${vm} cannot be recycled.`);
+        }
+        runConnectedCallback(vm);
+
+        if (vm.renderMode !== RenderMode.Light) {
+            // VM is not rendering in Light DOM, we can proceed and hydrate the slotted content.
+            // Note: for Light DOM, this is handled while hydrating the VM
+            hydrateChildrenHook(vnode.elm.childNodes, vnode.children, vm);
+        }
+
+        hydrateVM(vm);
+    },
+};
+
+function sameVnode(vnode1: VNode, vnode2: VNode): boolean {
+    return vnode1.key === vnode2.key && vnode1.sel === vnode2.sel;
+}
+
+function isVNode(vnode: any): vnode is VNode {
+    return vnode != null;
+}
 
 function observeElementChildNodes(elm: Element) {
     (elm as any).$domManual$ = true;
@@ -51,7 +338,18 @@ function setScopeTokenClassIfNecessary(elm: Element, owner: VM) {
     }
 }
 
-export function updateNodeHook(oldVnode: VNode, vnode: VNode) {
+function linkNodeToShadow(elm: Node, owner: VM) {
+    const { renderMode, shadowMode } = owner;
+
+    // TODO [#1164]: this should eventually be done by the polyfill directly
+    if (isSyntheticShadowDefined) {
+        if (shadowMode === ShadowMode.Synthetic || renderMode === RenderMode.Light) {
+            (elm as any)[KEY__SHADOW_RESOLVER] = getRenderRoot(owner)[KEY__SHADOW_RESOLVER];
+        }
+    }
+}
+
+function updateNodeHook(oldVnode: VNode, vnode: VNode) {
     const { elm, text } = vnode;
 
     if (oldVnode.text !== text) {
@@ -65,7 +363,7 @@ export function updateNodeHook(oldVnode: VNode, vnode: VNode) {
     }
 }
 
-export function insertNodeHook(vnode: VNode, parentNode: Node, referenceNode: Node | null) {
+function insertNode(vnode: VNode, parentNode: Node, referenceNode: Node | null) {
     if (process.env.NODE_ENV !== 'production') {
         unlockDomMutation();
     }
@@ -75,7 +373,7 @@ export function insertNodeHook(vnode: VNode, parentNode: Node, referenceNode: No
     }
 }
 
-export function removeNodeHook(vnode: VNode, parentNode: Node) {
+function removeNode(vnode: VNode, parentNode: Node) {
     if (process.env.NODE_ENV !== 'production') {
         unlockDomMutation();
     }
@@ -85,7 +383,7 @@ export function removeNodeHook(vnode: VNode, parentNode: Node) {
     }
 }
 
-export function patchElementPropsAndAttrs(oldVnode: VElement | null, vnode: VElement) {
+function patchElementPropsAndAttrs(oldVnode: VElement | null, vnode: VElement) {
     if (isNull(oldVnode)) {
         applyEventListeners(vnode);
         applyStaticClassAttribute(vnode);
@@ -100,16 +398,12 @@ export function patchElementPropsAndAttrs(oldVnode: VElement | null, vnode: VEle
     patchProps(oldVnode, vnode);
 }
 
-export const enum LWCDOMMode {
-    manual = 'manual',
-}
-
-export function hydrateElmHook(vnode: VElement) {
+function hydrateElmHook(vnode: VElement) {
     applyEventListeners(vnode);
     patchProps(null, vnode);
 }
 
-export function fallbackElmHook(elm: Element, vnode: VElement) {
+function fallbackElmHook(elm: Element, vnode: VElement) {
     const { owner } = vnode;
     setScopeTokenClassIfNecessary(elm, owner);
     if (owner.shadowMode === ShadowMode.Synthetic) {
@@ -120,7 +414,7 @@ export function fallbackElmHook(elm: Element, vnode: VElement) {
         if (
             !isUndefined(context) &&
             !isUndefined(context.lwc) &&
-            context.lwc.dom === LWCDOMMode.manual
+            context.lwc.dom === LwcDomMode.manual
         ) {
             // this element will now accept any manual content inserted into it
             observeElementChildNodes(elm);
@@ -136,7 +430,7 @@ export function fallbackElmHook(elm: Element, vnode: VElement) {
         const isPortal =
             !isUndefined(context) &&
             !isUndefined(context.lwc) &&
-            context.lwc.dom === LWCDOMMode.manual;
+            context.lwc.dom === LwcDomMode.manual;
         const isLight = owner.renderMode === RenderMode.Light;
         patchElementWithRestrictions(elm, { isPortal, isLight });
     }
@@ -150,7 +444,7 @@ export function patchChildren(parent: ParentNode, oldCh: VNodes, newCh: VNodes) 
     }
 }
 
-export function allocateChildrenHook(vnode: VCustomElement, vm: VM) {
+function allocateChildrenHook(vnode: VCustomElement, vm: VM) {
     // A component with slots will re-render because:
     // 1- There is a change of the internal state.
     // 2- There is a change on the external api (ex: slots)
@@ -176,7 +470,7 @@ export function allocateChildrenHook(vnode: VCustomElement, vm: VM) {
     }
 }
 
-export function createViewModelHook(elm: HTMLElement, vnode: VCustomElement) {
+function createViewModelHook(elm: HTMLElement, vnode: VCustomElement) {
     if (!isUndefined(getAssociatedVMIfPresent(elm))) {
         // There is a possibility that a custom element is registered under tagName,
         // in which case, the initialization is already carry on, and there is nothing else
@@ -205,7 +499,7 @@ export function createViewModelHook(elm: HTMLElement, vnode: VCustomElement) {
     }
 }
 
-export function createChildrenHook(vnode: VElement) {
+function createChildrenHook(vnode: VElement) {
     const { elm, children } = vnode;
     for (let j = 0; j < children.length; ++j) {
         const ch = children[j];
@@ -390,7 +684,7 @@ export function hydrateChildrenHook(elmChildren: NodeListOf<ChildNode>, children
     }
 }
 
-export function removeElmHook(vnode: VElement) {
+function removeChildren(vnode: VElement) {
     // this method only needs to search on child vnodes from template
     // to trigger the remove hook just in case some of those children
     // are custom elements.
@@ -460,4 +754,186 @@ export function markAsDynamicChildren(children: VNodes) {
 
 function hasDynamicChildren(children: VNodes): boolean {
     return FromIteration.has(children);
+}
+
+function createKeyToOldIdx(
+    children: VNodes,
+    beginIdx: number,
+    endIdx: number
+): Record<Key, number> {
+    const map: Record<Key, number> = {};
+
+    // TODO [#1637]: simplify this by assuming that all vnodes has keys
+    for (let j = beginIdx; j <= endIdx; ++j) {
+        const ch = children[j];
+        if (isVNode(ch)) {
+            const { key } = ch;
+            if (key !== undefined) {
+                map[key] = j;
+            }
+        }
+    }
+    return map;
+}
+
+function addVnodes(
+    parentElm: Node,
+    before: Node | null,
+    vnodes: VNodes,
+    startIdx: number,
+    endIdx: number
+) {
+    for (; startIdx <= endIdx; ++startIdx) {
+        const ch = vnodes[startIdx];
+        if (isVNode(ch)) {
+            ch.hook.create(ch);
+            ch.hook.insert(ch, parentElm, before);
+        }
+    }
+}
+
+function removeVnodes(parentElm: Node, vnodes: VNodes, startIdx: number, endIdx: number): void {
+    for (; startIdx <= endIdx; ++startIdx) {
+        const ch = vnodes[startIdx];
+        // text nodes do not have logic associated to them
+        if (isVNode(ch)) {
+            ch.hook.remove(ch, parentElm);
+        }
+    }
+}
+
+function updateDynamicChildren(parentElm: Node, oldCh: VNodes, newCh: VNodes) {
+    let oldStartIdx = 0;
+    let newStartIdx = 0;
+    let oldEndIdx = oldCh.length - 1;
+    let oldStartVnode = oldCh[0];
+    let oldEndVnode = oldCh[oldEndIdx];
+    const newChEnd = newCh.length - 1;
+    let newEndIdx = newChEnd;
+    let newStartVnode = newCh[0];
+    let newEndVnode = newCh[newEndIdx];
+    let oldKeyToIdx: any;
+    let idxInOld: number;
+    let elmToMove: VNode | null | undefined;
+    let before: any;
+    while (oldStartIdx <= oldEndIdx && newStartIdx <= newEndIdx) {
+        if (!isVNode(oldStartVnode)) {
+            oldStartVnode = oldCh[++oldStartIdx]; // Vnode might have been moved left
+        } else if (!isVNode(oldEndVnode)) {
+            oldEndVnode = oldCh[--oldEndIdx];
+        } else if (!isVNode(newStartVnode)) {
+            newStartVnode = newCh[++newStartIdx];
+        } else if (!isVNode(newEndVnode)) {
+            newEndVnode = newCh[--newEndIdx];
+        } else if (sameVnode(oldStartVnode, newStartVnode)) {
+            patchVnode(oldStartVnode, newStartVnode);
+            oldStartVnode = oldCh[++oldStartIdx];
+            newStartVnode = newCh[++newStartIdx];
+        } else if (sameVnode(oldEndVnode, newEndVnode)) {
+            patchVnode(oldEndVnode, newEndVnode);
+            oldEndVnode = oldCh[--oldEndIdx];
+            newEndVnode = newCh[--newEndIdx];
+        } else if (sameVnode(oldStartVnode, newEndVnode)) {
+            // Vnode moved right
+            patchVnode(oldStartVnode, newEndVnode);
+            newEndVnode.hook.move(oldStartVnode, parentElm, nextSibling(oldEndVnode.elm!));
+            oldStartVnode = oldCh[++oldStartIdx];
+            newEndVnode = newCh[--newEndIdx];
+        } else if (sameVnode(oldEndVnode, newStartVnode)) {
+            // Vnode moved left
+            patchVnode(oldEndVnode, newStartVnode);
+            newStartVnode.hook.move(oldEndVnode, parentElm, oldStartVnode.elm!);
+            oldEndVnode = oldCh[--oldEndIdx];
+            newStartVnode = newCh[++newStartIdx];
+        } else {
+            if (oldKeyToIdx === undefined) {
+                oldKeyToIdx = createKeyToOldIdx(oldCh, oldStartIdx, oldEndIdx);
+            }
+            idxInOld = oldKeyToIdx[newStartVnode.key!];
+            if (isUndefined(idxInOld)) {
+                // New element
+                newStartVnode.hook.create(newStartVnode);
+                newStartVnode.hook.insert(newStartVnode, parentElm, oldStartVnode.elm!);
+                newStartVnode = newCh[++newStartIdx];
+            } else {
+                elmToMove = oldCh[idxInOld];
+                if (isVNode(elmToMove)) {
+                    if (elmToMove.sel !== newStartVnode.sel) {
+                        // New element
+                        newStartVnode.hook.create(newStartVnode);
+                        newStartVnode.hook.insert(newStartVnode, parentElm, oldStartVnode.elm!);
+                    } else {
+                        patchVnode(elmToMove, newStartVnode);
+                        oldCh[idxInOld] = undefined as any;
+                        newStartVnode.hook.move(elmToMove, parentElm, oldStartVnode.elm!);
+                    }
+                }
+                newStartVnode = newCh[++newStartIdx];
+            }
+        }
+    }
+    if (oldStartIdx <= oldEndIdx || newStartIdx <= newEndIdx) {
+        if (oldStartIdx > oldEndIdx) {
+            // There's some cases in which the sub array of vnodes to be inserted is followed by null(s) and an
+            // already processed vnode, in such cases the vnodes to be inserted should be before that processed vnode.
+            let i = newEndIdx;
+            let n;
+            do {
+                n = newCh[++i];
+            } while (!isVNode(n) && i < newChEnd);
+            before = isVNode(n) ? n.elm : null;
+            addVnodes(parentElm, before, newCh, newStartIdx, newEndIdx);
+        } else {
+            removeVnodes(parentElm, oldCh, oldStartIdx, oldEndIdx);
+        }
+    }
+}
+
+function updateStaticChildren(parentElm: Node, oldCh: VNodes, newCh: VNodes) {
+    const oldChLength = oldCh.length;
+    const newChLength = newCh.length;
+
+    if (oldChLength === 0) {
+        // the old list is empty, we can directly insert anything new
+        addVnodes(parentElm, null, newCh, 0, newChLength);
+        return;
+    }
+    if (newChLength === 0) {
+        // the old list is nonempty and the new list is empty so we can directly remove all old nodes
+        // this is the case in which the dynamic children of an if-directive should be removed
+        removeVnodes(parentElm, oldCh, 0, oldChLength);
+        return;
+    }
+    // if the old list is not empty, the new list MUST have the same
+    // amount of nodes, that's why we call this static children
+    let referenceElm: Node | null = null;
+    for (let i = newChLength - 1; i >= 0; i -= 1) {
+        const vnode = newCh[i];
+        const oldVNode = oldCh[i];
+        if (vnode !== oldVNode) {
+            if (isVNode(oldVNode)) {
+                if (isVNode(vnode)) {
+                    // both vnodes must be equivalent, and se just need to patch them
+                    patchVnode(oldVNode, vnode);
+                    referenceElm = vnode.elm!;
+                } else {
+                    // removing the old vnode since the new one is null
+                    oldVNode.hook.remove(oldVNode, parentElm);
+                }
+            } else if (isVNode(vnode)) {
+                // this condition is unnecessary
+                vnode.hook.create(vnode);
+                // insert the new node one since the old one is null
+                vnode.hook.insert(vnode, parentElm, referenceElm);
+                referenceElm = vnode.elm!;
+            }
+        }
+    }
+}
+
+function patchVnode(oldVnode: VNode, vnode: VNode) {
+    if (oldVnode !== vnode) {
+        vnode.elm = oldVnode.elm;
+        vnode.hook.update(oldVnode, vnode);
+    }
 }
