+++ packages/@lwc/engine-core/src/framework/modules/attrs.ts
@@ -4,16 +4,14 @@
  * SPDX-License-Identifier: MIT
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
  */
-import { isNull, isUndefined, StringCharCodeAt } from '@lwc/shared';
+import { isNull, isUndefined, StringCharCodeAt, XML_NAMESPACE, XLINK_NAMESPACE } from '@lwc/shared';
 
 import { setAttribute, removeAttribute } from '../../renderer';
-import { VElement } from '../../3rdparty/snabbdom/types';
 
 import { unlockAttribute, lockAttribute } from '../attributes';
 import { EmptyObject } from '../utils';
+import { VElement } from '../vnodes';
 
-const xlinkNS = 'http://www.w3.org/1999/xlink';
-const xmlNS = 'http://www.w3.org/XML/1998/namespace';
 const ColonCharCode = 58;
 
 export function patchAttributes(oldVnode: VElement | null, vnode: VElement) {
@@ -36,10 +34,10 @@ export function patchAttributes(oldVnode: VElement | null, vnode: VElement) {
             unlockAttribute(elm!, key);
             if (StringCharCodeAt.call(key, 3) === ColonCharCode) {
                 // Assume xml namespace
-                setAttribute(elm, key, cur as string, xmlNS);
+                setAttribute(elm, key, cur as string, XML_NAMESPACE);
             } else if (StringCharCodeAt.call(key, 5) === ColonCharCode) {
                 // Assume xlink namespace
-                setAttribute(elm, key, cur as string, xlinkNS);
+                setAttribute(elm, key, cur as string, XLINK_NAMESPACE);
             } else if (isNull(cur) || isUndefined(cur)) {
                 removeAttribute(elm, key);
             } else {
