+++ packages/@lwc/engine-core/src/framework/modules/static-style-attr.ts
@@ -7,7 +7,7 @@
 import { isUndefined } from '@lwc/shared';
 
 import { setCSSStyleProperty } from '../../renderer';
-import { VElement } from '../../3rdparty/snabbdom/types';
+import { VElement } from '../vnodes';
 
 // The HTML style property becomes the vnode.data.styleDecls object when defined as a string in the template.
 // The compiler takes care of transforming the inline style into an object. It's faster to set the
