+++ packages/@lwc/shared/src/namespaces.ts
@@ -1,0 +1,12 @@
+/*
+ * Copyright (c) 2022, salesforce.com, inc.
+ * All rights reserved.
+ * SPDX-License-Identifier: MIT
+ * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
+ */
+
+export const HTML_NAMESPACE = 'http://www.w3.org/1999/xhtml';
+export const XML_NAMESPACE = 'http://www.w3.org/XML/1998/namespace';
+export const SVG_NAMESPACE = 'http://www.w3.org/2000/svg';
+export const MATHML_NAMESPACE = 'http://www.w3.org/1998/Math/MathML';
+export const XLINK_NAMESPACE = 'http://www.w3.org/1999/xlink';
